moved app endpoint to TeamAPI
